Feat DIAL-10753 [Work In Progress]: Organize screens into dashboard and auth folders.
