+++ app/src/main/java/com/example/bottomnav/presentation/screens/dashboard/contacts/ContactDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.example.bottomnav.presentation.screens.account
+package com.example.bottomnav.presentation.screens.dashboard.contacts
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,8 +8,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun AccountScreen() {
+fun ContactDetailScreen() {
     Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text("Account Screen")
+        Text("Contact Detail Screen")
     }
 }
