+++ app/src/main/java/com/example/bottomnav/presentation/screens/dashboard/messages/MessageScreen.kt
@@ -1,4 +1,4 @@
-package com.example.bottomnav.presentation.screens.keypad
+package com.example.bottomnav.presentation.screens.dashboard.messages
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,8 +8,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun KeypadScreen() {
+fun MessageScreen() {
     Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text("Keypad Screen")
+        Text("Message Screen")
     }
 }
